Added Comment Functionality

+++ src/controllers/blogControllers.ts
@@ -7,6 +7,7 @@ import {
 	findBlogsByUserId,
 	updateBlogById,
 } from '@models/blogModel';
+import { findUserById } from '@models/userModel';
 import { Request, Response } from 'express';
 
 export const getAllBlogs = async (req: Request, res: Response) => {
@@ -39,6 +40,11 @@ export const getBlogsByUserId = async (req: Request, res: Response) => {
 	const userId = req.params.userId;
 
 	try {
+		const existingUser = await findUserById(userId);
+		if (!existingUser) {
+			return res.status(404).json('USER NOT FOUND');
+		}
+
 		const blogs = await findBlogsByUserId(userId);
 		return res.status(200).json(blogs);
 	} catch (error) {
@@ -47,17 +53,18 @@ export const getBlogsByUserId = async (req: Request, res: Response) => {
 };
 
 export const writeBlog = async (req: Request, res: Response) => {
-	const { title, content, thumbnail, userId } = req.body;
+	const { title, content, thumbnail } = req.body;
+	const { id } = req.user;
 
 	try {
-		const blog = await createBlog({ title, content, thumbnail, userId });
+		const blog = await createBlog({ title, content, thumbnail, userId: id });
 		return res.status(201).json(blog);
 	} catch (error) {
 		return res.status(500).json(error);
 	}
 };
 
-export const updateBlogDetails = async (req: Request, res: Response) => {
+export const updateBlog = async (req: Request, res: Response) => {
 	const blogId = req.params.blogId;
 	const updatedDetails: TBlog = req.body;
 
@@ -77,6 +84,11 @@ export const updateBlogDetails = async (req: Request, res: Response) => {
 export const deleteBlog = async (req: Request, res: Response) => {
 	const blogId = req.params.blogId;
 	try {
+		const existingBlog = await findBlogById(blogId);
+		if (!existingBlog) {
+			return res.status(404).json('BLOG NOT FOUND');
+		}
+
 		await deleteBlogById(blogId);
 		return res.status(200).json('BLOG DELETED');
 	} catch (error) {
