+++ src/controllers/commentControllers.ts
@@ -1,13 +1,16 @@
-import { updateBlogById } from '@models/blogModel';
+import { findBlogById, updateBlogById } from '@models/blogModel';
 import {
 	TComment,
-	createBlogComment,
+	createComment,
 	deleteCommentById,
 	findAllComments,
 	findCommentById,
 	findCommentsByBlogId,
 	findCommentsByUserId,
+	findRepliesByCommentId,
+	updateCommentById,
 } from '@models/commentModel';
+import { findUserById } from '@models/userModel';
 import { Request, Response } from 'express';
 
 export const getAllComments = async (req: Request, res: Response) => {
@@ -24,6 +27,10 @@ export const getCommentById = async (req: Request, res: Response) => {
 
 	try {
 		const comment = await findCommentById(commentId);
+		if (!comment) {
+			return res.status(404).json('COMMENT NOT FOUND');
+		}
+
 		return res.status(200).json(comment);
 	} catch (error) {
 		return res.status(500).json(error);
@@ -33,6 +40,11 @@ export const getCommentById = async (req: Request, res: Response) => {
 export const getCommentsByBlogId = async (req: Request, res: Response) => {
 	const blogId = req.params.blogId;
 	try {
+		const existingBlog = await findBlogById(blogId);
+		if (!existingBlog) {
+			return res.status(404).json('BLOG NOT FOUND');
+		}
+
 		const comment = await findCommentsByBlogId(blogId);
 		return res.status(200).json(comment);
 	} catch (error) {
@@ -43,6 +55,11 @@ export const getCommentsByBlogId = async (req: Request, res: Response) => {
 export const getCommentsByUserId = async (req: Request, res: Response) => {
 	const userId = req.params.userId;
 	try {
+		const existingUser = await findUserById(userId);
+		if (!existingUser) {
+			return res.status(404).json('USER NOT FOUND');
+		}
+
 		const comments = await findCommentsByUserId(userId);
 		return res.status(200).json(comments);
 	} catch (error) {
@@ -50,15 +67,74 @@ export const getCommentsByUserId = async (req: Request, res: Response) => {
 	}
 };
 
-export const writeComment = async (req: Request, res: Response) => {
-	const { content, blogId, userId } = req.body;
+export const getRepliesByCommentId = async (req: Request, res: Response) => {
+	const commentId = req.params.commentId;
+	try {
+		const existingComment = await findCommentById(commentId);
+		if (!existingComment) {
+			return res.status(404).json('COMMENT NOT FOUND');
+		}
+
+		const comment = await findRepliesByCommentId(commentId);
+		return res.status(200).json(comment);
+	} catch (error) {
+		return res.status(500).json(error);
+	}
+};
+
+export const writeBlogComment = async (req: Request, res: Response) => {
+	const { content, blogId } = req.body;
+	const { id } = req.user;
 
 	try {
-		const comment = await createBlogComment({
+		const existingBlog = await findBlogById(blogId);
+		if (!existingBlog) {
+			return res.status(404).json('BLOG NOT FOUND');
+		}
+
+		const comment = await createComment({
 			content,
 			blogId,
-			creatorId: userId,
+			creatorId: id,
 		});
+
+		const comments = existingBlog.comments;
+		comments.push(comment._id);
+
+		await updateBlogById(blogId, { comments: comments });
+
+		return res.status(200).json(comment);
+	} catch (error) {
+		return res.status(500).json(error);
+	}
+};
+
+export const writeReplyComment = async (req: Request, res: Response) => {
+	const { content, commentId } = req.body;
+	const { id } = req.user;
+
+	try {
+		const existingComment = await findCommentById(commentId);
+		if (!existingComment) {
+			return res.status(404).json('COMMENT NOT FOUND');
+		}
+
+		if (existingComment.creatorId == id) {
+			return res.status(404).json('CANNOT REPLY TO YOUR OWN COMMENT');
+		}
+
+		const comment = await createComment({
+			content,
+			creatorId: id,
+			blogId: existingComment.blogId,
+			parentCommentId: commentId,
+		});
+
+		const existingCommentReplies = existingComment.replies;
+		existingCommentReplies.push(comment._id);
+
+		await updateCommentById(commentId, { replies: existingCommentReplies });
+
 		return res.status(200).json(comment);
 	} catch (error) {
 		return res.status(500).json(error);
@@ -75,7 +151,7 @@ export const updateComment = async (req: Request, res: Response) => {
 			return res.status(404).json('COMMENT NOT FOUND');
 		}
 
-		const comment = await updateBlogById(commentId, updatedComment);
+		const comment = await updateCommentById(commentId, updatedComment);
 		return res.status(200).json(comment);
 	} catch (error) {
 		return res.status(500).json(error);
@@ -85,6 +161,11 @@ export const updateComment = async (req: Request, res: Response) => {
 export const deleteComment = async (req: Request, res: Response) => {
 	const commentId = req.params.commentId;
 	try {
+		const existingComment = await findCommentById(commentId);
+		if (!existingComment) {
+			return res.status(404).json('COMMENT NOT FOUND');
+		}
+
 		await deleteCommentById(commentId);
 		return res.status(200).json('COMMENT DELETED');
 	} catch (error) {
